backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart
moviepy==1.0.3
gTTS==2.5.0
numpy<2.0.0
Pillow==10.1.0
proglog
requests